+++ FE/src/components/calender/index.tsx
@@ -1,37 +1,21 @@
 import * as React from 'react';
+import CalenderPicker from 'components/calender/components/CalenderPicker';
+import { DateProvider, FilterType } from 'components/calender/context';
 
-import { getDateInfo, getCurrentMonthInfo, getMovedMonthInfo } from 'components/calender/utils/dataUtils';
-import Icon, { IconType } from 'components/calender/components/Icons';
-import * as Styled from 'components/calender/style';
-import CalenderList from './components/Calender';
-
-interface CalenderProps {
+export interface ICalenderProps {
   count: number;
   year: number;
   month: number;
   isSlider: boolean;
+  filter: FilterType;
 }
 
-const CalenderPicker = ({
-  count = 1,
-  year = new Date().getFullYear(),
-  month = new Date().getMonth() + 1,
-  isSlider = true,
-}: CalenderProps) => {
-  const getPrevMonthInfo = getMovedMonthInfo(year, month, 1);
-  const getNextMonthInfo = getMovedMonthInfo(year, month, -1);
-
+const CalenderApp = (props: ICalenderProps) => {
   return (
-    <Styled.CalenderWrapper count={count}>
-      {isSlider && (
-        <Styled.IconWrapper count={count}>
-          <Icon type={IconType.ARROW_LEFT} />
-          <Icon type={IconType.ARROW_RIGHT} />
-        </Styled.IconWrapper>
-      )}
-      <CalenderList count={count} year={year} month={month} />
-    </Styled.CalenderWrapper>
+    <DateProvider>
+      <CalenderPicker {...props} />
+    </DateProvider>
   );
 };
 
-export default CalenderPicker;
+export default CalenderApp;
